refactor: centering routes to avoid code duplicate and create name pattern
