+++ app/src/pages/ActivitiesPages/MyOfferedHelp/index.js
@@ -80,7 +80,7 @@ export default function HelpsFinished({ navigation }) {
                                     key={help._id}
                                     onPress={() =>
                                         navigation.navigate(
-                                            'MyOfferHelpDescription',
+                                            'myOfferHelpDescription',
                                             {
                                                 helpId: help._id,
                                                 routeId: 'HelpOffer',
